+++ backend/src/services/authentication.ts
@@ -2,16 +2,6 @@ import { TRPCError } from '@trpc/server'
 import { randomBytes } from 'crypto'
 import { getCache, setCache } from './cache.js'
 
-export const checkCredentials = (username: string, password: string) => {
-  const USERNAME = process.env.ADMIN_USERNAME
-  const PASSWORD = process.env.ADMIN_PASSWORD
-
-  if (!USERNAME || !PASSWORD) throw new Error('Username or password not set')
-
-  if (username === USERNAME && password === PASSWORD) return true
-  return false
-}
-
 export const createAuthorizedToken = async (username: string) => {
   const token = generateToken()
